backend/requirements.txt
# Core Framework
fastapi>=0.115.6
uvicorn[standard]>=0.30.6
pydantic>=2.9.2

# LangChain & AI Components
langchain>=0.2.14
langgraph>=0.2.26
langchain-openai>=0.1.23

# Database & Vector Store
redis>=5.0.8
qdrant-client>=1.11.3
sentence-transformers>=3.0.1

# Authentication
python-jose>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
httpx>=0.27.2
geopy>=2.4.1
orjson>=3.10.7
python-multipart>=0.0.9